Scroll to the element using _accessibilityElementWithLabel:...
If the element is not visible, the UIView will probably not exist and the
executionBlock will get the UIView that is visible and nearest to the
accessibility element in the hierarchy. This is not easy for the block to
predict and deal with, so use |_accessibilityElementWithLabel:| to make the
UIView visible.